RAM optimizations, some bitmaps placed to EXTRA ROM basic editing functions in generator icons placed into class integration with GIT (about firmware screen)

+++ Source/Framework/Classes.h
@@ -92,7 +92,8 @@ public:
 		bottom = max(bottom, rcUnion.bottom);
 	}
 
-	int left, top, right, bottom;
+	// RAM optimization int->short
+	short left, top, right, bottom;
 };
 
 template <class TYPE>
